Fix: Bad returned type of method getVolume in interface IVocalPlayer

+++ src/main/java/fr/pederobien/vocal/client/impl/AbstractPlayer.java
@@ -68,7 +68,7 @@ public abstract class AbstractPlayer implements IVocalPlayer {
 	}
 
 	@Override
-	public double getVolume() {
+	public float getVolume() {
 		return volume;
 	}
